Add presentational Viewing-menu (Timeline filter) demo story
Move the Playground surface data and pure derivation helpers out of
Timeline.stories.tsx into internal/timelinePlaygroundData so both the
Playground and the new filter demo share them. Behavior is preserved:
the Playground stories render identically.

Add Timeline.filter.features.stories.tsx: an ActionMenu multi-select
"Viewing" control that filters Timeline rows consumer-side. The opening
and closing lifecycle bookends stay pinned outside the predicate so the
timeline never renders empty regardless of toggles; the consumer owns
that guarantee, not the Timeline or control components.

Story and internal scaffolding only: no public API export, no changeset.
The authoritative Timeline taxonomy stays in github-ui.
