+++ packages/react/src/Timeline/Timeline.filter.features.stories.module.css
@@ -1,0 +1,4 @@
+/* Story-only styles for the Timeline "Viewing" filter demo. Not shipped. */
+.Toolbar {
+  margin-block-end: var(--base-size-16);
+}
