+++ packages/react/src/Timeline/Timeline.stories.module.css
@@ -52,16 +52,11 @@
 }
 
 /*
- * Timeline Playground: story-local styles for the representative event bodies and the empty
- * state rendered by the `TimelinePlayground` story (its controls are Storybook argTypes).
+ * Timeline Playground: story-local style for the empty state rendered by the
+ * `TimelinePlayground` story (its controls are Storybook argTypes).
  */
 
 .PlaygroundEmpty {
   max-width: 1012px;
   color: var(--fgColor-muted);
-}
-
-.Strong {
-  font-weight: var(--base-text-weight-semibold);
-  color: var(--fgColor-default);
 }
